pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "TransferMole backend"
authors = ["Ivan Loboda <loboda.ivan.y@gmail.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = {extras = ["standard"], version = "^0.30.6"}
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
deepdiff = "^8.0.1"
base58 = "^2.1.1"
stripe = "^8.7.0"
web3 = "^6.11.1"
redis = {extras = ["hiredis"], version = "^5.0.8"}
prometheus-fastapi-instrumentator = "^7.0.0"
cryptography = "^43.0.1"
cachetools = "^5.5.0"
python-telegram-bot = "^21.6"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
pydantic = {extras = ["mypy"], version = "^2.9.2"}
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
locust = "^2.31.6"
mypy = "^1.11.2"
types-psycopg2 = "^2.9.21.20240819"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
follow_imports = "normal"
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
ignore_missing_imports = false
disable_error_code = "import-untyped"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
